% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootml-package.R
\docType{package}
\name{bootml-package}
\alias{bootml-package}
\alias{bootml}
\title{bootml}
\description{
Bootstrap methods for selecting the tuning parameter for LASSO-type regression models
}
\author{
Jonas Striaukas (maintainer) \email{jonas.striaukas@gmail.com}
}
\keyword{package}
